[MPA-23] Implement Login UI
Add colors, fonts and and styles based on the current information available in Figma.

Implement base button and snackbarcomponents base on current information in Figma.

Prepare the app to include the setup of a darkmode and apply the correct colors according with the dark theme available.

Add logic to set the application to be edge to edge and setup the background and content colors of status and navigation bar.

Implement the authencation and empty home screen.